[MIG] adhoc_modules migration to 11.0 (#44)
* [ADD] adhoc_modules: update module version

* [ADD] adhoc_modules: remove shebang for python 3.x

* [ADD] adhoc_modules: update README.rst file

Follow our module template description

* [ADD] adhoc_modules: remove logger not used.

* [FIX] adhoc_modules: python 3.x reduce bultin

This one was removed, not use fnctools.reduce instead.

* [FIX] adhoc_modules: python 3.x http.server

Deprecated SimpleHTTPServer and BaseHTTPServer

* [ADD] adhoc_modules: remove old migration scripts.

* [FIX] adhoc_modules: missing Action To Do

* [FIX] adhoc_modules: future print_function import not needed

The print method is not used so this one is not really needed.

* [REF] adhoc_modules: fix misspelling error

* [FIX] adhoc_modules: Registry.get() deprecated

use Registry() directly instead.

* [ADD] adhoc_modules: add list of functionalities in README

* [pylint] Fix W7908 (missing-newline-extrafiles)

* [pylint] Fix C8108 (method-compute)

* [pylint] Fix C8109 (method-search)

* [pylint] Fix W8104 (api-one-deprecated)

Use api.depends properly for fuction fields.

* [pylint] Fix C8104 (class-camelcase)

* [pylint] Fix C8110 (method-inverse)

* [IMP] adhoc_modules: remove web_server for update

This was a way to show a webpage telling the user that we were updating the
instance. this one is not used anymore since the update of a customer database
is made by restart a rancher container: the customer system is down while the
update is in progress.

* [pylint] Fix R7980 (consider-merging-classes-inherited)

* [REM] adhoc_modules: auto reload functionality

This onw is to simulate the update of the module list check as to install from
the subscriptio. This one is not working and we also have a possible new way
to manage this with a new oca module web_ir_actions_act_view_reload